+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix
   - [`Current Folder`] Fix a issue that Current folder plugin's settings do not load #4326
   - [`Clipboard`] Fix a issue that Clipboard plugin cannot paste to active window #4328
+  - [`Wpm`] Fix a issue where WPM couldn't create script plugins #4330
 
 ## v2.0.0-beta.8 — 2026-01-10